feat : 트윗 좋아요 수 등 데이터 출력, 유저 페이지 내 작성했던 트윗 이력 조회 기능 추가

+++ src/main/kotlin/com/swj9707/twittercloneapiserver/v1/user/dto/UserResDTO.kt
@@ -25,12 +25,7 @@ class UserResDTO {
 
         data class UserProfile (
             val userProfile : UserDTO.Dto.TwitterUserProfile,
-            val countOfTweet : Long
-            //TODO
-            /*
-            * countOfFollower
-            * countOfFollowing
-            * */
+            val countOfTweet : Int
         )
 
         data class EditProfile(
